Rename form-input.input to form-input.text
