+++ react/__tests__/Form.test.tsx
@@ -3,7 +3,7 @@ import { render } from '@vtex/test-tools/react'
 
 import Form from '../Form'
 import FormInputCheckbox from '../FormInput-checkbox'
-import FormInput from '../FormInput-input'
+import FormText from '../FormInput-text'
 import FormFieldGroup from '../FormFieldGroup'
 
 test('should render full schema', () => {
@@ -23,8 +23,8 @@ test('should render full schema', () => {
 test('should render partial schema', () => {
   const { getAllByText } = render(
     <Form entity="asdf" schema="asdf">
-      <FormInput pointer="#/properties/firstName" />
-      <FormInput pointer="#/properties/lastName" />
+      <FormText pointer="#/properties/firstName" />
+      <FormText pointer="#/properties/lastName" />
       <FormFieldGroup pointer="#/properties/address" />
       <FormInputCheckbox pointer="#/properties/agreement" />
     </Form>
